Seven-segment-display-using-LPC1769-and-Free-RTOS-and-task-creation
A project in LPC Xpresso software is created for LPC 1769 to create the
Timer which can count up to on hour is made. This timing is shown in
seven segment display. The project uses Free RTOS to create and handle
tasks givng priority to each task as per the requirement.